separate panelSection and SelectionCardWrapper into separate components
